adding 'route paths' to Constants folder + fixing SideMenu activeRoute bug
