+++ src/lib/common/Constants.ts
@@ -1,10 +1,12 @@
 
-export const ROUTE_ACCOUNT = "/account";
-export const ROUTE_ACCOUNT_PROFILE = "/account/profile";
-export const ROUTE_ACCOUNT_ADDRESSES = "/account/addresses";
-export const ROUTE_ACCOUNT_ORDERS = "/account/orders";
-export const ROUTE_ACCOUNT_PAYMENTS = "/account/payments";
+export const ROUTE_ACCOUNT = "/account/";
+export const ROUTE_ACCOUNT_ACCESS = "/account/access/";
+export const ROUTE_ACCOUNT_ADDRESSES = "/account/addresses/";
+export const ROUTE_ACCOUNT_ORDERS = "/account/orders/";
+export const ROUTE_ACCOUNT_PAYMENTS = "/account/payments/";
 
 export const ROUTE_LOGIN = "/api/auth/signin";
 export const ROUTE_LOGOUT = "/api/auth/signout";
 // export const ROUTE_ACCOUNT_CART = "/account";
+
+export const ROUTE_PRODUCTS = "/products/"
