+++ src/pages/account/index.tsx
@@ -1,55 +1,98 @@
-
-import AccountWrapper from "@/components/layouts/AccountWrapper";
-import CenteredWrapper from "@/components/layouts/CenteredWrapper";
+import React, {useState} from 'react';
+import axios from 'axios'
 
 import { GetServerSidePropsContext } from "next";
 import { isLoginRequiredServer } from "@/lib/auth";
-
 import { fetchUserByEmail } from "@/lib/mongoDB/userQueries";
 import { UserPageProps } from "@/lib/common/Types";
-import { Box, Grid, Typography, Button, TextField } from "@mui/material";
+
+import AccountWrapper from "@/components/layouts/AccountWrapper";
+import CenteredWrapper from "@/components/layouts/CenteredWrapper";
 import { GridHeader } from "@/components/GridHeader";
-import { FormReadOnly } from "@/components/forms/FormReadOnly";
+import { Box, Typography, Button, Alert, AlertTitle } from "@mui/material";
 
-const Account = ({user, ...props}:UserPageProps) =>{
+import { FormReadOnly } from '@/components/forms/FormReadOnly';
+import { ProfileForm } from '@/components/forms/ProfileForm';
 
-    const {email, password, image} = user;
+import { ProfileFormData } from '@/lib/common/Types';
+import { AlertState, StatusOptions } from '@/lib/common/Types';
 
+
+const initAlert = {
+    status: StatusOptions.ERROR,
+    message: "",
+    isDisplay: false,
+}
+const alertTimer = 1000;
+
+const Profile = ({user, ...props}:UserPageProps) =>{
+
+    const [alert, setAlert] = useState<AlertState>(initAlert);
+    const [isEdit, setIsEdit] = useState(false);
+    const [userSnapshot, setUserSnapshot] = useState(user);
+
+    const {username, first_name, last_name, telephone} = userSnapshot;
     const data = [
-        {title: "Email", value: email},
-        {title: "Password", value: password.split("").map(pwd => '*')},
-        {title: "Profile Image", value: image},
+        {title: "Username", value: username},
+        {title: "First Name", value: first_name},
+        {title: "Last Name", value: last_name},
+        {title: "Telephone", value: telephone},
     ];
     const formReadOnlyElems = <FormReadOnly data={data} />
+
+    const handleSubmit = async (formData:ProfileFormData) => {
+        try {
+            const response = await axios.post('/api/forms/profile', {
+                id: user.id, ...formData
+            });
+
+            setUserSnapshot({...userSnapshot, ...formData});
+            setAlert({status: StatusOptions.SUCCESS, message: response.data.message, isDisplay: true});
+        } catch (error) {
+            console.log("🚀 ~ file: profile.tsx:40 ~ handleSubmit ~ error:", error)
+            setAlert({status: StatusOptions.ERROR, message: "Error updating profile.", isDisplay: true});
+        }
+
+        setIsEdit(false);
+    }
+
+    const formElem = <ProfileForm user={userSnapshot} handleSubmit={handleSubmit} />
+
+    const alertElem = alert.isDisplay? <Alert severity={`${alert.status}`}
+        onClose={ () => setAlert(initAlert) }
+    >
+        <AlertTitle>{alert.message}</AlertTitle>
+    </Alert> : <></>;
+
 
     return <>
         <AccountWrapper>
             <CenteredWrapper>
-                {/* <h1>Account Page</h1> */}
-
                 <Box>
+
                     <GridHeader >
                         <Typography component="span" variant="h5" ml={2} >
-                            Account Settings
+                            Profile
                         </Typography>
                         <Button variant="outlined" sx={{ marginRight: '0.5rem' }}
-                            // onClick={() => setIsEdit(!isEdit)}
+                            onClick={() => setIsEdit(!isEdit)}
                         >
-                            Edit
-                            {/* {!isEdit? "Edit" : "Cancel"} */}
+                            {!isEdit? "Edit" : "Cancel"}
                         </Button>
                     </GridHeader>
 
-                    {/* {isEdit? formElem : formReadOnlyElems} */}
-                    {formReadOnlyElems}
+                    {alertElem}
+
+                    {isEdit? formElem : formReadOnlyElems}
 
                 </Box>
-
-
             </CenteredWrapper>
         </AccountWrapper>
     </>
 }
+
+
+
 
 export async function getServerSideProps(context:GetServerSidePropsContext) {
     const check = await isLoginRequiredServer(context);
@@ -80,4 +123,4 @@ export async function getServerSideProps(context:GetServerSidePropsContext) {
     };
 }
 
-export default Account;
+export default Profile;
